Fix images api route to return proper headers
